Separate setting S3 bucket name from production configuration check
The bucket_name method was conflating two concerns: (a) setting the S3
bucket name based on an environment variable; and (b) checking that the
S3 bucket name is set in a production environment.

The motivation to do this is to make it possible to override the latter
check for the publishing-e2e tests so that they can use the fake S3
functionality in a Rails production environment.