Allow use of fake S3 functionality in production
The motivation behind this is specifically for the publishing-e2e-tests
project. The idea is to run the app in the Rails production environment
in those tests in order to make them more realistic. This is intended to
be an interim measure until a more realistic fake S3 service (e.g.
Minio/Localstack) can be setup.

As highlighted in #338, we're a bit uncomfortable with this approach,
because it makes it possible to accidentally configure the app to use
the fake S3 functionality in the real production environment which would
be very undesirable. However, I've mitigated this risk by making the
name of the relevant environment variable very explicit.

/cc @RyanMacG, @adrianclay, @kevindew