Introduce uploaded & not_uploaded states
I'm planning to use the new methods which these states introduce in the
SaveToCloudStorageWorker to update the state of an asset when an attempt
is make to upload the asset to cloud storage.

It feels as if it makes sense for any assets other than unscanned or
infected ones to be marked as uploaded or not_uploaded, e.g. I can
imagine we might want to allow an already uploaded asset to be
re-uploaded, and we definitely should allow a not_uploaded asset to have
the upload retried (i.e. the default behaviour of Sidekiq for jobs where
an exception has been raised).

Note that I'm planning to use the "bang" version of the state transtion
methods so that the jobs fail fast if the state transition is invalid.