Introduce uploaded state to Asset
I'm planning to use the upload_success! method (which ths state
introduces) in the SaveToCloudStorageWorker to update the state of an
asset when an attempt to upload the asset to cloud storage succeeds.

Note that I'm planning to use the "bang" version of the state transtion
methods so that the jobs fail fast if the state transition is invalid.

Note also that I've constrained the state machine only to allow
transitions into the new "uploaded" state from the "clean" state as
suggested by @chrisroos.