Change asset state on successful/failed upload to cloud storage
The CloudStorage::ObjectUploadFailedError exception will be raised if
the upload fails.

If the upload succeeds, the asset state is changed to "uploaded";
whereas if it fails, the state is changed to "not_uploaded".

If the upload fails, the CloudStorage::ObjectUploadFailedError exception
is still raised in order to trigger the Sidekiq retry mechanism.

For the moment I'm continuing to treat assets in the two new states as
if they were in the "clean" state. This means there is no external
behaviour change at this point. This is why I've made use of the shared
examples in the two media controller specs. I had to change the logic in
MediaController#asset_servable? to pass these new specs, but the logic
in the WhitehallMediaController passed them without needing any changes.

I plan to change the behaviour so that assets which are clean, but have
not yet been uploaded are treated the same as unscanned assets in the
two media controllers.

Note that only *new* assets will end up in either of the new states. I
plan to introduce a Rake task to move existing clean assets into one or
other of the two new states in a subsequent commit.