Change asset state on successful upload to cloud storage
The S3Storage::ObjectUploadFailedError exception will be raised if the
upload fails.

If the upload succeeds, the asset state is changed to "uploaded";
whereas if it fails, the state is not changed i.e. remains set to
"clean".

If the upload fails, the S3Storage::ObjectUploadFailedError exception
will trigger the Sidekiq retry mechanism.

For the moment I'm continuing to treat assets in the new state as if
they were in the "clean" state. This means there is no external
behaviour change. This is why I've made use of the shared examples in
the two media controller specs. I had to change the logic in
MediaController#asset_servable? to pass these new specs, but the logic
in the WhitehallMediaController passed them without needing any changes.

In subsequent commits, I plan to change the behaviour so that assets
which are clean, but have not yet been uploaded are treated the same as
unscanned assets in the two media controllers.

Note that only *new* assets will end up in the new state. I plan to
introduce a Rake task to move existing clean assets into the new state
in a subsequent commit.