Add Rake task to set uploaded/not_uploaded state for clean assets
The state is set to uploaded or not_uploaded depending on whether a
corresponding object exists on S3.

For now the app treats assets in clean, uploaded, or not_uploaded states
the same, but once this Rake task has been run, the plan is to have
different behaviour for these states.

Note that I've re-introduced the AssetProcessor, because clean assets
comprise the majority of the assets and so the Rake task will take a
while to run and I want to avoid the MongoDB cursor problems I ran into
a while ago with a similar long-running Rake task.

Similarly I'm using a Sidekiq worker on the low_priority queue to
do the actual work.