Explicitly handle Aws::S3::MultipartUploadError in S3Storage#save
According to the documentation for Aws::S3::Object#upload_file [1] the
method may raise this exception if an object is being uploaded in parts,
and the upload can not be completed.

Previously this exception was not handled and therefore the
S3Storage#save method was already failing fast and any Sidekiq job being
processed by SaveToCloudStorageWorker would be retried automatically.

By handling this exception and treating it the same way as a `false`
return value we can simplify the API of S3Storage and make it easier for
clients to use. Note that since we raise an app-specific exception, the
Sidekiq job retries should still kick in.

[1]:
http://docs.aws.amazon.com/sdkforruby/api/Aws/S3/Object.html#upload_file-instance_method