Make integration spec more representative
This example was a pretty good end-to-end check that everything was
working as expected. However, it was never extended when we started
uploading assets to S3 and proxying asset requests to S3.

I've added some extra steps to one of the examples to test the uploading
to (fake) S3 and proxying of asset requests to (fake) S3.

Note the asset request to the Rails app results in a 200 OK response
whether the asset is in the clean or uploaded state. However, since all
asset requests are now proxied to S3, when the asset is in the clean
state the proxied request to S3 results in a 404 Not Found, because the
object does not yet exist on S3. This behaviour can be observed in the
raising of an ActionController::RoutingError exception by the fake S3
service within the new section of this example.

I plan to address this in a subsequent PR and I'm hoping that including
these changes in this PR will make the subsequent changes more obvious.
See #281 for more details.

While there is an argument to rename this spec to something more
generic, I'm happy to leave it unchanged for the moment.