Prevent change in routes leaking out of fake S3 routing spec
Before this was fixed I was seeing the newly extended
`spec/requests/virus_scanning_psec.rb` failing intermittently depending
on whether it was run before or after this fake S3 routing spec. The
problem was that the former spec relies on the fake S3 route existing,
but without the changes introduced in this commit, sometimes the fake S3
route disappeared from the application routes.