Respond with 404 Not Found for Mainstream assets not yet uploaded to S3
Previously the Rails app was responding with a 200 OK and setting the
X-Accel-Redirect header to point to a non-existent object on S3. This
still resulted in the end-user seeing a 404 Not Found, but since the
Rails app "knows" when it has uploaded the object to S3, it seems better
to have the Rails app respond with the 404 Not Found and avoid a request
to S3.