Redirect to placeholders for Whitehall assets not yet uploaded to S3
Previously the Rails app was responding with a 200 OK and setting the
X-Accel-Redirect header to point to a non-existent object on S3. This
potentially resulted in the end-user seeing a 404 Not Found, albeit only
for the short time between an asset being virus scanned successfully and
being uploaded to S3. This change means that the app will now respond to
requests for clean assets in the same way as it does for unscanned
assets; it's only when the asset is in the uploaded state that the app
will now respond with a 200 OK.

See #281 for details.