Remove unnecessary optimisation in VirusScanner
The VirusScanner class is only ever instantiated in the VirusScanWorker
and in that case, the #clean method is only called once on that
instance. Thus this optimisation is only ever exercised in the
VirusScanner spec example which I have removed.

My motivation for making this change is that I want to convert
VirusScanner into a stateless service to bring it into line with
S3Storage and services in other GOV.UK apps. This change is a step in
that direction.