Raise exception when VirusScanner#scan detects a virus
Previously this method was returning `false` if a virus was detected.
I'm converting this class into a stateless service and so I want to
avoid having to store `virus_info` when a virus is detected. We can do
this by raising the exception and handling it in the VirusScanWorker.
Handling the exception in the worker is important, because we want to
avoid triggering the Sidekiq retry mechanism.

Note that the `virus_info` string is now passed as the exception message
and since we send the original exception to GovukError.notify, there's
no longer any need to include it in the `extra` Hash.