Delete underlying file when upload to S3 succeeds
Since #328 the app hasn't been capable of instructing Nginx to serve
assets directly from the NFS mount and since alphagov/govuk-puppet#6971
Nginx hasn't been configured to allow assets to be served directly from
the NFS mount.

This means that the only reason we need to store files on NFS is
currently so that the Sidekiq workers (which may be running on different
machines to that which handled the upload request) have access to the
files.

This in turn means that it is safe to remove the underlying file once
the asset has been scanned for viruses and uploaded to S3.

In order to achieve this I've introduced a new state transition callback
which calls `Asset#remove_file!`. The latter seems to be the recommended way
to remove the underlying file via the Carrierwave API and results in
`Asset#file` becoming blank which seems to make sense to me.

I've had to restrict the validation which checks for the presence of the
file attribute in order to ensure that an uploaded asset remains valid.
I've also had to prevent the `store_metadata` callback from running when
the asset is in the uploaded state, because it relies on the underlying
file in order to e.g. compute the ETag from the file metadata. I think
disabling the latter should be OK, because it should only need to be run
on Asset creation or update, i.e. when a new file is specified.

Once this commit has been deployed, the underlying files for assets
should only exist on NFS temporarily, i.e. until they have been scanned
for viruses and uploaded to S3. This means we should then be in a
position to delete all the underlying files for existing assets in the
uploaded state and free up a significant amount of space on NFS.