Extract update_asset script logic and add test coverage
Note that I've attempted as unchanged as possible while making
sufficient changes to allow for sensible testing.

In the course of doing this, I realised that there was a flaw in the
logic and the `filename` usage message would never be displayed, because
an IndexError would be raised when `ARGV.fetch(1)` was called when ARGV
is a single element array. I've fixed this and simplified the code in
doing so.