Percent encode public_url_path in asset JSON
We've found a number of Whitehall assets whose filenames (and therefore
their `legacy_url_path`s) contain non-ascii characters. We can safely
store these non-ascii characters in the database but need to escape them
when constructing the `file_url` in the JSON representation of an asset.
This addresses the bug described in issue #395.

I initially used Ruby's `URI.escape` method but Rubocop doesn't like
that as it's been marked as obsolete.

After this change we can upload and request a Whitehall asset whose
legacy_url_path contains a non-ascii character:

```
$ echo `date` > whitehall-assét.txt

$ curl http://asset-manager.dev.gov.uk/whitehall_assets \
  --form "asset[file]=@whitehall-assét.txt" \
  --form "asset[legacy_url_path]=/government/uploads/path/to/whitehall-assét.txt"
{
  "_response_info":{"status":"created"},
  "id":"http://asset-manager.dev.gov.uk/assets/5a562f05759b745bcb180de6",
  "name":"whitehall-ass_t.txt",
  "content_type":"text/plain",
  "file_url":"http://assets-origin.dev.gov.uk/government/uploads/path/to/whitehall-ass%C3%A9t.txt",
  "state":"unscanned"
}

$ curl
http://asset-manager.dev.gov.uk/government/uploads/path/to/whitehall-ass%C3%A9t.txt
Wed Jan 10 15:18:02 UTC 2018
```

Note that the version of CarrierWave we're using (0.11.x) doesn't
support non-ascii characters in uploaded filenames which is why the
`name` in the example JSON above has replaced the 'é' with an
underscore.