Support Whitehall's ConsultationResponseFormUploader ext whitelist
This ensure we're setting the correct content type for assets allowed by
Whitehall's `ConsultationResponseFormUploader`[1].

I used `Mime::Type.lookup_by_extension` in the Whitehall project to
generate the expected mime types used in these tests.

Note the change to the "and the extension is not a recognised mime type"
test. I had to update the extension now that '.doc' is a known type.

[1]: https://github.com/alphagov/whitehall/blob/4040195b57c9ef9a25e6580a08ca7e8c68673ce0/app/uploaders/consultation_response_form_uploader.rb