Remove authentication for private virtual host
Although in govuk-puppet the `PRIVATE_ASSET_MANAGER_HOST` environment
variable is set to `private-asset-manager.${app_domain}` [1] and
`private-asset-manager` is added as an alias to the `asset-manager`
virtual host [2], the Asset Manager app never supplies URLs using this
hostname to publishing apps and I haven't found any evidence that any of
the publishing apps manually construct URLs using this hostname. So as
far as I can tell this functionality is not currently (and possibly has
never been) used.

We're about to add functionality to the Asset Manager to protect draft
assets behind SSO and this unused functionality just confused matters.

I plan to remove the related code from govuk-puppet in a PR on that
repo.

[1]:
https://github.com/alphagov/govuk-puppet/blob/9a4ad116d8b495690415405e6ee3b0663959d06d/modules/govuk/manifests/apps/asset_manager.pp#L69-L74
[2]:
https://github.com/alphagov/govuk-puppet/blob/9a4ad116d8b495690415405e6ee3b0663959d06d/modules/govuk/manifests/apps/asset_manager.pp#L172