Remove access-limited functionality for Mainstream assets
At some point it was possible to mark an asset as access-limited.
However, since the upgrade to Rails v4 [1] this will have been prevented
by Rails' Strong Parameter functionality [2]. This won't have been
noticed, because there are no specs testing that it's possible to supply
the `access_limited` or `organisation_slug` parameters to the API
endpoints for creating or updating an asset. If either of these
parameters are supplied to these API endpoints, the parameters are
stripped from the attributes used in the mass assignment and the
a warning like the following is logged:

    Unpermitted parameters: :access_limited, :organisation_slug

I confirmed that there are currently no assets in production with either
of these fields set by asking for the following queries to be run in the
Asset Manager Rails console:

    irb> Asset.unscoped.where(access_limited: nil).count
    => 0
    irb> Asset.unscoped.where(access_limited: false).count
    => 1100658
    irb> Asset.unscoped.where(access_limited: true).count
    => 0

    irb> Asset.unscoped.where(organisation_slug: nil).count
    => 1100658
    irb> Asset.unscoped.where(:organisation_slug.ne =>
    nil).count
    => 0

We're about to add functionality to Asset Manager to protect draft
assetes using SSO and this unused functionality will just confuse
matters.

[1]: https://github.com/alphagov/asset-manager/pull/42
[2]:
http://guides.rubyonrails.org/v5.1.4/action_controller_overview.html#strong-parameters