Remove unused Asset#access_limited? & #organisation_slug
As mentioned in a previous commit, there are no assets with these fields
set in production and there is no longer any code which uses them, so
it's safe to remove them.

I've added a Rake task so we can *ensure* that the fieldis are removed
from all existing assets, even though this is probably already the case.