Add Asset#draft Boolean field defaulting to false
We want to be able to mark assets as draft in order to be able to
protect draft Whitehall attachments behind SSO.

Although at the moment, we only need to do this for Whitehall
attachments, there is a longer term plan for other publishing
applications to make use of this. That's why I've chosen to add the
field to Asset and not WhitehallAsset.

It feels as if the draft state of an asset is orthogonal to Asset#state.
This is why I've chosen to make it a separate field rather than adding
an extra state to the Asset#state state machine.

For the moment, the majority of assets will be created as published
assets. That's why I've chosen to default the value of Asset#draft to
`false`, even though this might change in the future.