Make GovukConfiguration available from AssetManager accessor
This brings the GovukConfiguration into line with other configuration
objects. My motivation for doing this is that I'm planning to add a new
method to GovukConfiguration and this will mean I can easily get access
to it from anywhere in the codebase.

In adding this, I noticed the redundant `AssetManager.app_host`
accessor, but I plan to remove that in a separate PR.