Add GovukConfiguration#draft_assets_host method
This makes use of the new `Plek#external_url_for` method which gives us
access to the external base URL for a service. This is useful, because
in AWS environments the internal and external base URLs are different
for a given service.

We're going to need this to decide whether a request needs to be
authenticated.

In making this change, I wondered whether the implementation of
`GovukConfiguration#app_host` should be using `Plek` rather than
directly accessing env vars, but I plan to address that in a separate
PR.