Standardize code for redirecting to draft-assets host
The previous implementation was non-standard and therefore confusing. In
particular, the
`BaseMediaController#redirected_to_draft_assets_host_for?` method looked
like a query method, but was actually a command method, i.e. it had side
effects which might have been surprising. By splitting this method into
a query method and a command method, we can avoid this confusion at the
expense of a little duplication.