Extract WhitehallAsset.from_params to reduce duplication
I'm about to add a third use of this logic in a new
WhitehallAssetsController#update action and so it seems like a sensible
time to DRY this up.

It's made slightly awkward by the need for the optional `path_prefix`
parameter, but I still think it's a worthwhile change. We might be able
to remove this complication by changing the route for
WhitehallMediaController#download so that the `path` param includes the
whole path and by adding a constraint to ensure it starts with
`/government/uploads`. However, I'm going to leave that for another day!