Make id param lookup idiomatic & consistent in AssetsController
I think it's more idiomatic Rails to just lookup the `id` from the
params using `ActionController::Parameters#[]` and rely on the call to
`Asset.find` failing, rather than calling
`ActionController::Parameters#fetch` which raises a
`ActionController::ParameterMissing` exception.

The `show` action was already using this approach and there are no
tests requiring the original behaviour, so I'm happy to make this
change.