Remove duplicate examples from AssetsController spec
This whole `context` block already exists in the `describe` block for
'POST create'.

I think there might've been an intention to use it as the basis for
testing what happens if the `update` action is called with invalid asset
attributes. However, the only validation we could conceivably use to
trigger such a failure is the validation for the presence of Asset#file.
However, this is an attribute added using CarrierWave and it doesn't
appear to behave like a normal Rails attribute in that when you call
`Asset#update_attributes` with `file: nil` it does not set the `file`
attribute to `nil` and so the model is still valid and
`Asset#update_attributes` returns `true`.

So in the end, I've elected just to remove the duplicate examples and
not attempt to add the ones that I think are missing.