Make SaveToCloudStorageWorker job idempotent
If the asset has already been uploaded, we shouldn't attempt to upload
it again, because the underlying file may have been removed by the
`after_transition to: :uploaded` callback.

My motivation for adding this is that on a couple of occasions I've seen
this job apparently fail, because the underlying file no longer exists.
And on at least one occasion the state of the asset has been changed to
'uploaded' and the asset appears to have been saved to the cloud. This
suggests that the job may have executed more than once. If that is
indeed what happened, this change should avoid that happening again.

Also the Sidekiq documentation recommends ensuring jobs are idempotent
[1].

[1]:
https://github.com/mperham/sidekiq/wiki/Best-Practices#2-make-your-job-idempotent-and-transactional