Use bang version of state machine event methods
Previously we were using the non-bang versions of the methods which
return false if they fail. However, we were ignoring the return value
and so they will just have silently failed.

The bang version of the methods will raise a
StateMachines:InvalidTransition exception if the transition is not
allowed. We're not expecting this to happen and so raising an exception
seems like a sensible thing to do. It also brings this worker into line
with the SaveToCloudStorageWorker.

I contemplated adding some examples to test this behaviour, but I
noticed we're not doing so in the SaveToCloudStorageWorker spec and in
any case, it just seems like good practice to use this version of the
methods. So I'm not going to bother.

Using the bang version in specs also just seems like good practice, i.e.
fail fast if an invalid transition is attempted.