Add constraints to Asset state machine transitions
This means that it's now only possible to transition from 'unscanned' to
'clean' or 'infected'; it's not possible to get to either of the latter
states from any state other than 'unscanned'. While I can imagine it
_might_ be useful to allow other transitions in the future, these
constraints reflect the current application's use of the state machine.

Constraining the state machine as much as possible makes the code
easier to reason about.