Only attempt virus scan if asset is in unscanned state
This aims to make the VirusScanWorker job idempotent.

My motivation for adding this is that on a couple of occasions I've seen
this job apparently fail, because the underlying file no longer exists.
As with the changes to SaveToCloudStorageWorker earlier in this PR, my
hypothesis is that this was caused by the job being executed more than
once for the same asset. If that is indeed what happened, this change
should avoid that happening again.

Also the Sidekiq documentation recommends ensuring jobs are idempotent
[1].

[1]:
https://github.com/mperham/sidekiq/wiki/Best-Practices#2-make-your-job-idempotent-and-transactional