Always set max-age to 0 for API asset show endpoints
This logic became out-of-date when we introduced the uploaded state in
this PR [1]. I considered only updating the logic to fix this, i.e. by
setting cache expiry to 0 for uploaded or infected assets (vs scanned or
infected assets).

However, I realised that since we introduced the `draft` status of an
asset in this PR [2] and included it in the JSON response, it no longer
makes sense to set the cache expiry to 0 for uploaded assets, because
their `draft` status may change.

Since very few assets are marked as infected, few (if any) of the
publishing apps use this endpoint, and it isn't a very "expensive"
request, so I decided it was better to simplify the app by always
setting the cache expiry to 0.

Note that gds-api-adapters gem used by the publishing apps does respect
the Cache-Control header [3] and so it is important to set the header
appropriately.

[1]: https://github.com/alphagov/asset-manager/pull/364
[2]: https://github.com/alphagov/asset-manager/pull/435
[3]:
https://github.com/alphagov/gds-api-adapters/blob/5bdb71743ca94f11cad2756b37ab150a847ed53c/lib/gds_api/json_client.rb#L204-L241