Allow location of clamscan to be configured by environment variable
In the end to end tests environment we want to be able to disable
the virus scan easily as it can take more than 60 seconds to complete
a scan when several requests are made to it at once.

By allowing the path to be passed in we can give it /bin/true when
wanting to disable the scan.