Add basic access_limiting functionality for mainstream assets
We want to be able to restrict access to draft assets to a subset of
users. This is only a requirement for Whitehall assets in the short term
but I think that enabling it for all assets will make Asset Manager
easier to understand.

I'm basing this access limitation in Asset Manager on a cut-down version
of the equivalent functionality in the Content Store; notably I'm not
storing the `auth_bypass_ids` that it's possible to set on a
`ContentItem` in the Content Store. Whitehall doesn't appear to set any
`auth_bypass_ids` when publishing to the Content Store (in fact, it
might only be Publisher that does set them) so I don't think we need to
either.

The `access_limited` attribute should either by an empty array (meaning
that the asset is accessible to everyone) or an array of user UIDs
(meaning that the asset is only accessible to users whose UID appear in
the list). This mirrors the behaviour in the Content Store.