Add Asset#redirect_uri and allow read/write via API
A request for a Whitehall attachment which is not publicly accessible
and is associated with an unpublished edition results in a 302 Temporary
Redirect to the unpublished document page, which will typically display
an explanation for why the document has been unpublished.

We're going to need to replicate this behaviour for Whitehall
attachments when they are served from Asset Manager, so we need a way
for Whitehall to tell Asset Manager about the redirect URL. This change
makes this possible.

I've chosen to add the new `redirect_url` field to Asset, rather than
just Whitehall asset, because it's simpler to do that and I can't see
why it wouldn't be generally useful behaviour for any asset.