Redirect requests for Whitehall assets with redirect_url
The plan is for Whitehall to set Asset#redirect_url for an attachment if
its parent document is unpublished.

This relates to the logic in Whitehall's AttachmentController#fail [1]
which determines whether the attachment's parent document has been
unpublished. In Asset Manager the concept of a "draft" asset means that
requests for it must come via the draft-assets host and be associated
with an authenticated user. Thus to avoid a double redirect and (more
significantly) unnecessary authentication, an asset with a redirect URL
should not be marked as draft.

Notes:

1. This redirect is intended to be used for non-asset URLs. More
specifically in the case of an unpublished document, the user should be
redirected to the original public URL of the relevant attachment's
parent document. This will be set by the Whitehall app.

2. I've added the new redirect logic *before* the placeholder redirect
logic in WhitehallMediaController#download to match the fact that
Whitehall's AttachmentController#fail gives this redirect priority over
the placeholder redirect.

3. I've only added the redirect logic for Whitehall assets for now,
since they're the only ones we actually need, but it will almost
certainly make sense to provide this functionality for all assets.

4. I believe there is a flaw in the logic in
AttachmentVisibility#unpublished_edition as described in point 2 in this
issue [2]. For the moment, it seems simplest to reproduce the flawed
behaviour.

[1]:
https://github.com/alphagov/whitehall/blob/64bb168edf383fc62a7190f614f6822df9dff4a7/app/controllers/attachments_controller.rb#L43-L44
[2]: https://github.com/alphagov/asset-manager/issues/479