Migration to set size for existing assets
Assets that have already been uploaded to Asset Manager do not have
their size field set. This commit adds a rake task to set the size
field for all of these assets (a previous commit ensures that the
`size` gets set on upload for all new assets).

We don't have the file available on disk to `stat` and set the size as
we now delete the file from NFS as soon as it has been uploaded to
S3. I considered downloading the file again as part of this migration,
but that would likely be slow. It may also have been possible to use
the S3 API to fetch the file size, but again this would require a
possibly slow network request. In the end I have decided to derive the
size from the hex-encoded final part of the `etag` which has already
been set on upload for all existing assets.

The Asset#size= method has been protected in an earlier commit, so the
`uploaded_asset_without_size` factory has to use `send` to create the
appropriate test data.

This commit can be reverted once the migration has been run on all
environments.