Use worker for asset size migration
In 16e323e we added a rake task to set the `size` field for all
existing assets. When we ran the task in integration however, we
estimated that it would take around 12 hours to complete. This runs
the risk of the rake task being killed and leaving the database in a
somewhat inconsistent state.

This commit moves the work of the rake task into a worker. Hopefully
this will improve the run time as the workers run with concurrency,
and the rake task that queues up the jobs will complete more quickly.