Upgrade aws-sdk from v2.10.105 -> v2.11.3, i.e. latest 2.x
I loosened the constraint on the aws-sdk gem in the Gemfile to allow
any 2.x version and then ran the following command:

    bundle update --conservative aws-sdk

This has upgraded to the latest 2.x version. I've looked through the
changelog [1] and it looks as if all the changes are new features, so
this should be a safe upgrade, although it would be sensible to test the
app in integration, i.e. with a real S3 connection.

As per the review on this PR [2], I don't want to upgrade to v3.x,
because it includes breaking changes.

[1]: https://github.com/aws/aws-sdk-ruby/blob/version-2/CHANGELOG.md
[2]: https://github.com/alphagov/asset-manager/pull/352