Add Asset#replacement association
When a Whitehall attachment is "replaced", requests for its URL result
in a redirect to the attachment which replaced it. We're going to need
to implement this functionality in Asset Manager in order to serve
Whitehall attachments without changing the externally visible behaviour.
This is a first step in that direction.

Note that as of Rails v5, a belongs_to association is required by
default. Hence why I've had to specify the `optional: true` option.

I didn't see any point in implementing the inverse relation at this
point, because it's not obvious we're going to need it.

Note that some of the model examples I've added to the Asset spec may
not be needed, but I wanted to make sure I understood how this
association was persisted by Mongoid.