Include replacement_id in JSON response for GET API endpoint
If asset has a replacement then include its ID in the JSON response. The
convention seems to be to include such attributes in the response by
default; the only exception being if there is some kind of security
consideration. In this case it seems fine to include the new attribute.