Allow replacement_id to be specified via the API
This commit makes it possible to specify a replacement_id when creating
or updating an asset.

Although the most likely use case is when *updating* an existing asset,
I've also allowed it when creating a new asset for completeness.

Although the immediate need for this if only for Whitehall assets, I've
also allowed it for Mainstream assets, because it seems generally
useful. Note that I suspect we could probably replace the
MediaController#redirect_to_current_filename behaviour with this more
general mechanism at some point.